+++ fuel-asm/src/panic_instruction.rs
@@ -13,6 +13,7 @@ use crate::{
 #[derive(fuel_types::canonical::Deserialize, fuel_types::canonical::Serialize)]
 /// Describe a panic reason with the instruction that generated it
 pub struct PanicInstruction {
+    #[canonical(skip)]
     reason: PanicReason,
     instruction: RawInstruction,
 }
@@ -112,5 +113,24 @@ impl From<Word> for PanicInstruction {
             reason,
             instruction,
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::op;
+    use fuel_types::canonical::Serialize;
+
+    #[test]
+    fn canonical_serialization_ignores_panic_reason() {
+        let revert_panic_instruction =
+            PanicInstruction::error(PanicReason::Revert, op::noop().into());
+        let out_of_gas_panic_instruction =
+            PanicInstruction::error(PanicReason::OutOfGas, op::noop().into());
+        assert_eq!(
+            revert_panic_instruction.to_bytes(),
+            out_of_gas_panic_instruction.to_bytes()
+        );
     }
 }